+++ src/meta/raft-store/src/sm_v002/mod.rs
@@ -15,12 +15,21 @@
 pub mod leveled_store;
 pub(in crate::sm_v002) mod marked;
 #[allow(clippy::module_inception)]
-pub mod sm_v002;
-pub mod snapshot_store;
-pub mod snapshot_view_v002;
-pub mod writer_v002;
+mod sm_v002;
+mod snapshot_store;
+mod snapshot_view_v002;
+mod writer_v002;
+
+mod importer;
 
 #[cfg(test)]
 mod sm_v002_test;
 #[cfg(test)]
 mod snapshot_view_v002_test;
+
+pub use importer::Importer;
+pub use sm_v002::SMV002;
+pub use snapshot_store::SnapshotStoreError;
+pub use snapshot_store::SnapshotStoreV002;
+pub use snapshot_view_v002::SnapshotViewV002;
+pub use writer_v002::WriterV002;
