+++ src/meta/raft-store/src/sm_v002/sm_v002_test.rs
@@ -18,7 +18,7 @@ use pretty_assertions::assert_eq;
 
 use crate::sm_v002::leveled_store::map_api::MapApi;
 use crate::sm_v002::marked::Marked;
-use crate::sm_v002::sm_v002::SMV002;
+use crate::sm_v002::SMV002;
 use crate::state_machine::ExpireKey;
 
 #[test]
